+++ core/ecosystem/pypi/const.go
@@ -1,8 +1,9 @@
 package pypi
 
 const (
-	EcosystemName = "PyPI"
-	PyProjectToml = "pyproject.toml"
-	PoetryLock    = "poetry.lock"
-	PdmLock       = "pdm.lock"
+	EcosystemLanguage = "Python"
+	EcosystemName     = "PyPI"
+	PyProjectToml     = "pyproject.toml"
+	PoetryLock        = "poetry.lock"
+	PdmLock           = "pdm.lock"
 )